Make it compile with all feature combinations
This mostly consisted of splitting the live and non live stream implementations into two separate files & gating them behind a feature
