+++ Cargo.toml
@@ -35,7 +35,7 @@ thiserror = "1.0.40"
 derive_more = "0.99.17"
 derivative = "2.2.0"
 once_cell = "1.17.1"
-tokio = { version = "1.26.0", optional = true }
+tokio = { version = "1.26.0", default-features = false, features = ["sync"] }
 rand = "0.8.5"
 reqwest-middleware = "0.2.2"
 reqwest-retry = "0.2.2"
@@ -52,13 +52,13 @@ tokio = { version = "1.26.0", features = ["full"] }
 [features]
 default = ["search", "live", "default-tls"]
 live = [
-    "tokio/rt",
-    "tokio/rt-multi-thread",
     "tokio/time",
-    "tokio/sync",
     "tokio/process",
 ]
-blocking = []
+blocking = [
+    "tokio/rt",
+    "tokio/rt-multi-thread"
+]
 search = []
 default-tls = [
     "reqwest/default-tls",
