+++ src/stream/mod.rs
@@ -1,8 +1,13 @@
 mod encryption;
 mod hashable_byte_range;
-mod media_format;
 mod remote_data;
-mod segment;
 mod streams;
 
-pub use streams::{LiveStream, LiveStreamOptions, NonLiveStream, NonLiveStreamOptions, Stream};
+#[cfg(feature = "live")]
+mod media_format;
+#[cfg(feature = "live")]
+mod segment;
+
+#[cfg(feature = "live")]
+pub use streams::{LiveStream, LiveStreamOptions};
+pub use streams::{NonLiveStream, NonLiveStreamOptions, Stream};
